dev(alias): removed old alias test
